read args struct change
