+++ vm/general_struct.c
@@ -17,7 +17,7 @@ t_general		*gen_init(void)
 	gen->players = NULL;
 	gen->champ_num = 0;
 	gen->line = NULL;
-	ft_bzero(gen->n_flag, MAX_PLAYERS + 1);
-	ft_bzero(gen->no_flag, MAX_PLAYERS + 1);
+//	ft_bzero(gen->n_flag, MAX_PLAYERS + 1);
+//	ft_bzero(gen->no_flag, MAX_PLAYERS + 1);
 	return(gen);
 }
