+++ vm/vm.h
@@ -40,9 +40,11 @@ void		ft_usage(void);
 int			ft_error(char *message);
 void		read_args(int ar, char **av, t_general *gen);
 int			is_flag(char **args, size_t *i, t_general *gen);
+void		is_n_flag(char **args, size_t *i, t_general *gen);
 int			is_valid_champ(char *arg, t_general *gen, size_t j);
 int			is_champ(char *arg);
 int			validate_champ(unsigned char *line, int i);
+int			is_valid_num(char *arg, char flag);
 t_general	*gen_init(void);
 void		read_players(char **av, t_general *gen);
 void		write_player(t_general *gen, size_t j);
